remove webhook code
